fixing the create SQL files

+++ ddl/create_book.sql
@@ -1,10 +1,11 @@
+# Book table
 CREATE TABLE book (
-  ISBN13   INT          NOT NULL,
-  title    VARCHAR(200) NOT NULL,
-  year     DATE         NOT NULL,
-  category VARCHAR(200) NOT NULL,
-  pname    VARCHAR(200) NOT NULL,
-  price    DECIMAL      NOT NULL,
+  ISBN13   BIGINT         NOT NULL,
+  title    VARCHAR(200)   NOT NULL,
+  year     DATE           NOT NULL,
+  category VARCHAR(200)   NOT NULL,
+  pname    VARCHAR(200)   NOT NULL,
+  price    DECIMAL(10, 2) NOT NULL,
   PRIMARY KEY pk_book (ISBN13),
   FOREIGN KEY fk_book (pname)
   REFERENCES publisher (pname)
