+++ ddl/create_people.sql
@@ -1,7 +1,8 @@
+# People table
 CREATE TABLE people (
   name      VARCHAR(200) NOT NULL,
   address   VARCHAR(200) NOT NULL,
-  telephone INT          NOT NULL,
+  telephone BIGINT       NOT NULL,
   email     VARCHAR(200) NOT NULL,
   PRIMARY KEY pk_people(name, address)
 );
